// ERP-CustomerSide/src/app/models/mownerModel.ts
export interface mownerModel {

    id: number;

    compNo: number;

    vesselName: string;

    description: string;

    tasks: number;

    openedDate: string;

    dueDate: string;

    closedDate: string;

    remarks: string;

    subject: string;

    human: boolean;

    system: boolean;

    material: boolean;

    status: number;

    docNo: string;

    extendedDate: string;

}
  